+++ src/main/scala/com/sfxcode/sapphire/javafx/stage/StageValues.scala
@@ -1,0 +1,5 @@
+package com.sfxcode.sapphire.javafx.stage
+
+object StageValues extends Enumeration {
+  val Development, Production, SystemTest, Staging, UnitTest = Value
+}